Fix ConnectionFactory._create_read_write_connection
Changes how the initial read_write connection is constructed to use the
write_config and then apply a read_connection.  This fixes a bug where
a config containing `read` and `write`keys is passed directly into the
connection, which results in an error.

Compare this change to the Laravel analog found at
Illuminate\Database\Connectors\ConnectionFactory::createReadWriteConnection() (line 75)