Adds an option to add columns with time zone
Adds an option to create time related columns like time, timestamp and
datetime with the option "with time zone" instead of just without it.
So we wouldn't break BC, the default is still false.
This is done by passing a with_time_zone=True as a parameter when
creating columns.