Add @recoverable decorator to retry operations upon connection error
Similar to the `@run` decorator, this decorator causes a wrapped
method to be retried in the case that the method throws an exception
which appears to be caused by a lost connection.

While the `@run` decorator is limited to wrapping methods with a
particular signature (including `query` and `bindings` params, the
`@recoverable` decorator is more generally applicable.